requirements.txt
certifi==2020.6.20
llvmlite==0.38.0
mouse==0.7.1
numba==0.55.1
numpy==1.21.5
Pillow==9.0.1
pynput==1.7.6
PySide2==5.15.2.1
pywin32==303
shiboken2==5.15.2.1
six==1.16.0
wincertstore==0.2
winput==1.2.2
